Make fetchMore API in BatchPagingIterator less error prone
Previously the `BatchPagingIterator` used `fetchMore` to request more
data, which was then supposed to be supplied via `loadComplete`.

This changes the communication API so that `fetchMore` returns a future
which will contain the result. Due to that:

 - It becomes impossible to call `loadComplete` early without having
 called `fetchMore` first. Removing a possible error case.

 - `PagingIterator` usage is further encapsulated within the
 `BatchPagingIterator`

 - It is easier to write a unit test for the `BatchPagingIterator` with
 a paged source.

The `CumulativePageBucketReceiver` could probably be reworked further to
make better use of the changed API, but I tried to keep the changes
there minimal for this commit.